fix(db): cast runtime_ticks param to bigint in ProbeOnPlay update
CASE WHEN $2 > 0 让 PostgreSQL 把 $2 推断为 int4，导致回写
runtime_ticks（100ns tick，超 ~214.7s 即超 int4 上限）时报
"unable to encode ... into binary format for int4"。显式 ::bigint
修复 ProbeOnPlay 播放回写溢出。

+++ internal/repository/background_task_repository.go
@@ -609,7 +609,7 @@ func (r *BackgroundTaskRepository) UpdateProbeMediaVersion(ctx context.Context, 
 	_, err := r.pool.Exec(ctx,
 		`UPDATE media_versions
 		 SET mediainfo = $1,
-		     runtime_ticks = CASE WHEN $2 > 0 THEN $2 ELSE runtime_ticks END,
+		     runtime_ticks = CASE WHEN $2::bigint > 0 THEN $2::bigint ELSE runtime_ticks END,
 		     bitrate = COALESCE($3, bitrate),
 		     size = COALESCE($4, size),
 		     chapters = $6
